extract Card.jsx
